+++ lib/profitable/processors/braintree_processor.rb
@@ -12,7 +12,7 @@ module Profitable
         interval = data['billing_period_unit']
         interval_count = data['billing_period_frequency'] || 1
 
-        normalize_to_monthly(amount * quantity, interval, interval_count)
+        normalize_to_monthly(amount.to_f * quantity, interval, interval_count)
       end
     end
   end